Up to version 0.7 with plugin support - easy to create your own plugins
